% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables.R
\name{addVariables}
\alias{addVariables}
\title{addVariables}
\usage{
addVariables(outputdir = c())
}
\arguments{
\item{outputdir}{...}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
addVariables
}
