% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_diary.R
\name{convert_diary}
\alias{convert_diary}
\title{convert_diary}
\usage{
convert_diary(activitylog = c())
}
\arguments{
\item{activitylog}{...}
}
\value{
no object is returned, only a new file is created in the same folder as the activitylog
}
\description{
convert_diary
}
