% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sleeplog.R
\name{convert_sleeplog}
\alias{convert_sleeplog}
\title{convert_sleeplog}
\usage{
convert_sleeplog(sleeplog = c(), part2resultsfile = c())
}
\arguments{
\item{sleeplog}{...}

\item{part2resultsfile}{...}
}
\value{
no object is returned, only a new file is created in the same folder as the activitylog
}
\description{
convert_sleeplog
}
