% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gt3x_to_csv.R
\name{gt3x_to_csv}
\alias{gt3x_to_csv}
\title{gt3x_to_csv}
\usage{
gt3x_to_csv(path = c(), outpath = c(), gzip = T)
}
\arguments{
\item{path}{full path to file(s)}

\item{outpath}{directory where converted files should be stored (if not supplied it will be in same location}

\item{gzip}{Boolean to indicate whether output file should be gunzipped}
}
\value{
no object is returned, only a new file is created
}
\description{
gt3x_to_csv
}
