% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_folder_paths.R
\name{obtain_folder_paths}
\alias{obtain_folder_paths}
\title{obtain_folder_paths}
\usage{
obtain_folder_paths()
}
\value{
no object is returned, GGIR writes all its outputs to files
}
\description{
obtain_folder_paths
}
