% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optain_folder_paths.R
\name{optain_folder_paths}
\alias{optain_folder_paths}
\title{optain_folder_paths}
\usage{
optain_folder_paths()
}
\value{
no object is returned, GGIR writes all its outputs to files
}
\description{
optain_folder_paths
}
