% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGGIR.R
\name{runGGIR}
\alias{runGGIR}
\title{runGGIR}
\usage{
runGGIR(
  datadir = c(),
  outputdir = c(),
  mode = c(),
  do.report = c(),
  overwrite = FALSE
)
}
\arguments{
\item{datadir}{...}

\item{outputdir}{...}

\item{mode}{...}

\item{do.report}{...}

\item{overwrite}{...}
}
\value{
no object is returned, GGIR writes all its outputs to files
}
\description{
runGGIR
}
