% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables5.R
\name{addVariables5}
\alias{addVariables5}
\title{addVariables5}
\usage{
addVariables5(outputdir = c(), sep = ",")
}
\arguments{
\item{outputdir}{Path to output directory generated by GGIR}

\item{sep}{separator used by csv files stored by GGIR}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
addVariables5
}
