% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGGIR.R
\name{runGGIR}
\alias{runGGIR}
\title{runGGIR}
\usage{
runGGIR(datadir = c(), outputdir = c(), activitylog = c(), mode = c())
}
\arguments{
\item{datadir}{...}

\item{outputdir}{...}

\item{activitylog}{...}

\item{mode}{...}
}
\value{
no object is returned, GGIR writes all its outputs to files
}
\description{
runGGIR
}
