% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGGIR.R
\name{runGGIR}
\alias{runGGIR}
\title{runGGIR}
\usage{
runGGIR(
  datadir = c(),
  outputdir = c(),
  mode = c(),
  do.report = c(),
  overwrite = FALSE,
  do.visual = FALSE,
  visualreport = FALSE,
  acc.metric = "BFEN",
  chunksize = 1
)
}
\arguments{
\item{datadir}{...}

\item{outputdir}{...}

\item{mode}{...}

\item{do.report}{...}

\item{overwrite}{...}

\item{do.visual}{...}

\item{visualreport}{...}

\item{acc.metric}{...}

\item{chunksize}{...}
}
\value{
no object is returned, GGIR writes all its outputs to files
}
\description{
runGGIR
}
