% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGGIR.R
\name{runGGIR}
\alias{runGGIR}
\title{runGGIR}
\usage{
runGGIR(
  datadir = c(),
  outputdir = c(),
  mode = c(),
  do.report = c(),
  overwrite = FALSE,
  do.visual = FALSE,
  visualreport = FALSE,
  acc.metric = "BFEN",
  chunksize = 1,
  loglocation = c(),
  do.parallel = TRUE,
  testbatch = FALSE,
  sleeplogidnum = TRUE,
  hrs.del.start = 0,
  hrs.del.end = 0,
  sleepwindowType = "TimeInBed",
  sep = ","
)
}
\arguments{
\item{datadir}{GGIR argument, see document R package GGIR}

\item{outputdir}{GGIR argument, see document R package GGIR}

\item{mode}{GGIR argument, see document R package GGIR}

\item{do.report}{GGIR argument, see document R package GGIR}

\item{overwrite}{GGIR argument, see document R package GGIR}

\item{do.visual}{GGIR argument, see document R package GGIR}

\item{visualreport}{GGIR argument, see document R package GGIR}

\item{acc.metric}{GGIR argument, see document R package GGIR}

\item{chunksize}{GGIR argument, see document R package GGIR}

\item{loglocation}{GGIR argument, see document R package GGIR}

\item{do.parallel}{GGIR argument, see document R package GGIR}

\item{testbatch}{If set to TRUE the function will only process the first three files}

\item{sleeplogidnum}{GGIR argument, see document R package GGIR}

\item{hrs.del.start}{GGIR argument, see document R package GGIR}

\item{hrs.del.end}{GGIR argument, see document R package GGIR}

\item{sleepwindowType}{GGIR argument, see document R package GGIR}

\item{sep}{separator used by csv files stored by GGIR}
}
\value{
no object is returned, GGIR writes all its outputs to files
}
\description{
runGGIR
}
