% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\title{summarise}
\usage{
summarise(outputdir, part5_summary, Nmostrecent = 10, model_threshold = 75)
}
\arguments{
\item{outputdir}{...}

\item{part5_summary}{...}

\item{Nmostrecent}{...}

\item{model_threshold}{...}
}
\value{
no object is returned, only a summary is printed to the screen
}
\description{
summarise
}
