% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\title{summarise}
\usage{
summarise(
  outputdir,
  part5_summary,
  model_threshold = 75,
  referentiewaarden,
  sleepwindowType = "SPT",
  MVPAdefinition = "MVPA_E5S_B10M80\%_T100_BFEN_0-24hr",
  threshold_wrist = 0.5,
  sep = sep
)
}
\arguments{
\item{outputdir}{ath to output directory generated by GGIR}

\item{part5_summary}{Content of GGIR part 5 personsummary file}

\item{model_threshold}{Acceleration threshold corresponding to the model, which can be derived as negative of the first coefficient divided by the second coefficient.}

\item{referentiewaarden}{Mean and standard deviation ENMO_fullRecordingMean in a reference population.}

\item{sleepwindowType}{See, GGIR documentation.}

\item{MVPAdefinition}{MVPA variable name from the GGIR part2 daysummary output}

\item{threshold_wrist}{Threshold to classify active or passive}

\item{sep}{separator used by csv files stored by GGIR}
}
\value{
no object is returned, only a summary is printed to the screen
}
\description{
summarise
}
